requirements.txt
seleniumbase
selenium
flask
supabase
python-dotenv
requests
